consistent signature for GTOmax_cache_size

+++ pyscf/lib/pbc/fill_ints_sr.c
@@ -33,7 +33,7 @@
 #define OF_CMPLX        2
 
 int GTOmax_shell_dim(int *ao_loc, int *shls_slice, int ncenter);
-int GTOmax_cache_size(int (*intor)(), int *shls_slice, int ncenter,
+size_t GTOmax_cache_size(int (*intor)(), int *shls_slice, int ncenter,
                       int *atm, int natm, int *bas, int nbas, double *env);
 
 static double get_dsqure(double *ri, double *rj)
